feat: work on log lists
