+++ src/data.js
@@ -166,17 +166,272 @@ const data = {
       balance: 48569,
       theme: "dark",
       favoriteAccounts: [],
-      logs: []
+      logs: [
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+      ]
     },
     enterprise: {
       owner: "Arthur Popov",
       balance: 485321,
-      logs: []
+      logs: [
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+      ]
     },
     offshore: {
       owner: "Quentin Cooper",
       balance: 4123547,
-      logs: []
+      logs: [
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+        {
+          label: "Dépot",
+          amount: 15000,
+          date: "2022-04-19T15:00",
+          reference: "Failygames",
+          type: "operation",
+        },
+        {
+          label: "Retrait",
+          amount: -5000,
+          date: "2022-04-15T15:00",
+          reference: "American Restaurant Company",
+          type: "operation",
+        },
+        {
+          label: "Transfert",
+          amount: -8000,
+          date: "2022-04-14T15:00",
+          reference: "Owen Chapman",
+          type: "transfer",
+        },
+      ]
     }
   }
 }
