--- src/scripts/log-list.js
@@ -1,26 +1,0 @@
-class LogList {
-  constructor(logs, HtmlElement) {
-    this.logs = logs;
-    this.HtmlElement = HtmlElement;
-  }
-
-  addLog(log) {
-
-  }
-
-  deleteLog(log) {
-
-  }
-
-  createLog(log) {
-
-  }
-
-  createList() {
-
-  }
-
-  resetList() {
-
-  }
-}
